+++ ocdsextensionregistry/exceptions.py
@@ -2,19 +2,19 @@ class OCDSExtensionRegistryError(Exception):
     """Base class for exceptions from within this package"""
 
 
-class DoesNotExist(OCDSExtensionRegistryError):
+class DoesNotExist(OCDSExtensionRegistryError):  # noqa: N818
     """Raised if an object wasn't found for the given parameters"""
 
 
-class MissingExtensionMetadata(OCDSExtensionRegistryError):
+class MissingExtensionMetadata(OCDSExtensionRegistryError):  # noqa: N818
     """Raised if a method call requires extensions metadata, with which the extension registry was not initialized"""
 
 
-class NotAvailableInBulk(OCDSExtensionRegistryError):
+class NotAvailableInBulk(OCDSExtensionRegistryError):  # noqa: N818
     """Raised if files are required to be available in bulk, but are not"""
 
 
-class UnknownLatestVersion(OCDSExtensionRegistryError):
+class UnknownLatestVersion(OCDSExtensionRegistryError):  # noqa: N818
     """Raised if the latest version of an extension can't be determined"""