+++ tests/test_codelist_code.py
@@ -1,5 +1,4 @@
 import pytest
-
 from ocdsextensionregistry import CodelistCode