+++ tests/test_util.py
@@ -1,5 +1,4 @@
 import pytest
-
 from ocdsextensionregistry import ExtensionVersion
 from ocdsextensionregistry.exceptions import UnknownLatestVersion
 from ocdsextensionregistry.util import get_latest_version
@@ -7,7 +6,7 @@ from ocdsextensionregistry.util import get_latest_version
 
 def test_get_latest_version_one():
     versions = [
-        ExtensionVersion(arguments(**{'Version': '1'})),
+        ExtensionVersion(arguments(Version='1')),
     ]
 
     assert get_latest_version(versions).version == '1'
@@ -15,9 +14,9 @@ def test_get_latest_version_one():
 
 def test_get_latest_version_master():
     versions = [
-        ExtensionVersion(arguments(**{'Version': '1'})),
-        ExtensionVersion(arguments(**{'Date': None})),
-        ExtensionVersion(arguments(**{'Date': '1000-01-01', 'Version': 'master'})),
+        ExtensionVersion(arguments(Version='1')),
+        ExtensionVersion(arguments(Date=None)),
+        ExtensionVersion(arguments(Date='1000-01-01', Version='master')),
     ]
 
     assert get_latest_version(versions).version == 'master'
@@ -25,9 +24,9 @@ def test_get_latest_version_master():
 
 def test_get_latest_version_default():
     versions = [
-        ExtensionVersion(arguments(**{'Version': '1'})),
-        ExtensionVersion(arguments(**{'Date': None})),
-        ExtensionVersion(arguments(**{'Date': '1000-01-01', 'Version': '1.1'})),
+        ExtensionVersion(arguments(Version='1')),
+        ExtensionVersion(arguments(Date=None)),
+        ExtensionVersion(arguments(Date='1000-01-01', Version='1.1')),
     ]
 
     assert get_latest_version(versions).version == '1.1'
@@ -35,9 +34,9 @@ def test_get_latest_version_default():
 
 def test_get_latest_version_dated():
     versions = [
-        ExtensionVersion(arguments(**{'Version': '1'})),
-        ExtensionVersion(arguments(**{'Date': None})),
-        ExtensionVersion(arguments(**{'Date': '1000-01-01'})),
+        ExtensionVersion(arguments(Version='1')),
+        ExtensionVersion(arguments(Date=None)),
+        ExtensionVersion(arguments(Date='1000-01-01')),
     ]
 
     assert get_latest_version(versions).version == '1'
@@ -45,8 +44,8 @@ def test_get_latest_version_dated():
 
 def test_get_latest_version_dateless():
     versions = [
-        ExtensionVersion(arguments(**{'Date': None})),
-        ExtensionVersion(arguments(**{'Date': None, 'Version': '1'})),
+        ExtensionVersion(arguments(Date=None)),
+        ExtensionVersion(arguments(Date=None, Version='1')),
     ]
 
     with pytest.raises(UnknownLatestVersion) as excinfo:
@@ -61,7 +60,7 @@ def arguments(**kwargs):
         'Date': '2019-02-25',
         'Version': 'v1.1.4',
         'Base URL': 'https://raw.githubusercontent.com/open-contracting-extensions/ocds_location_extension/v1.1.4/',
-        'Download URL': 'https://api.github.com/repos/open-contracting-extensions/ocds_location_extension/zipball/v1.1.4',  # noqa: E501
+        'Download URL': 'https://api.github.com/repos/open-contracting-extensions/ocds_location_extension/zipball/v1.1.4',
     }
 
     data.update(kwargs)